Update Client.java
ok so I modified the client in the sense that the username and password of the person can easily be obtained and sent to the server. However, the server must also be modified to be able to read the login message object, and change its status to either success of failure, so the client can act accordingly.